chore: removed redundant files and functionality and updated docs
