+++ libs/webcomponents-nx/package.json
@@ -2,6 +2,5 @@
   "name": "@ui5/webcomponents-nx",
   "version": "0.2.3-rc.0",
   "main": "src/index.js",
-  "generators": "./generators.json",
   "executors": "./executors.json"
 }
