+++ libs/webcomponents-nx/src/executors/prepare/prepareOptions.ts
@@ -2,9 +2,5 @@ import {SchematicsOptions} from "./schematicsOptions";
 
 export interface PrepareOptions {
   schematics?: SchematicsOptions;
-  packageJsonPath?: string;
-  versionsOverrides?: {
-    projectVersion?: string;
-  };
   distPath: string;
 }
