--- libs/webcomponents-nx/src/generators/sync/index.ts
@@ -1,22 +1,0 @@
-import { formatFiles, readProjectConfiguration, Tree } from '@nx/devkit';
-import {
-  fsCommit,
-  NxFsImplementation,
-} from '@ui5/webcomponents-wrapper-fs-commit';
-import { wrapper, WrapperConfig } from '@ui5/webcomponents-wrapper';
-
-export default async function generateComponents(
-  tree: Tree,
-  options: WrapperConfig<unknown> & { project: string }
-) {
-  const fsAdapter = new NxFsImplementation(tree);
-  const projectConfiguration = readProjectConfiguration(tree, options.project);
-  const basePath = projectConfiguration.sourceRoot;
-  await wrapper({
-    commit: fsCommit(fsAdapter, basePath),
-    ...options,
-  });
-  await formatFiles(tree);
-  // eslint-disable-next-line @typescript-eslint/no-empty-function
-  return () => {};
-}
