--- libs/webcomponents-nx/src/generators/sync/schema.json
@@ -1,20 +1,0 @@
-{
-  "$schema": "http://json-schema.org/schema",
-  "$id": "Sync",
-  "title": "",
-  "type": "object",
-  "properties": {
-    "project": {
-      "type": "string",
-      "description": "",
-      "$default": {
-        "$source": "argv",
-        "index": 0
-      },
-      "x-prompt": "What name would you like to use?"
-    }
-  },
-  "required": [
-    "project"
-  ]
-}
