--- libs/webcomponents-nx/src/generators/webcomponents-nx/files/src/index.ts__template__
@@ -1,1 +1,0 @@
-const variable = "<%= projectName %>";
