--- libs/webcomponents-nx/src/generators/webcomponents-nx/schema.d.ts
@@ -1,5 +1,0 @@
-export interface WebcomponentsNxGeneratorSchema {
-  name: string;
-  tags?: string;
-  directory?: string;
-}
