--- libs/webcomponents-nx/src/generators/webcomponents-nx/schema.json
@@ -1,29 +1,0 @@
-{
-  "$schema": "http://json-schema.org/schema",
-  "$id": "WebcomponentsNx",
-  "title": "",
-  "type": "object",
-  "properties": {
-    "name": {
-      "type": "string",
-      "description": "",
-      "$default": {
-        "$source": "argv",
-        "index": 0
-      },
-      "x-prompt": "What name would you like to use?"
-    },
-    "tags": {
-      "type": "string",
-      "description": "Add tags to the project (used for linting)",
-      "alias": "t"
-    },
-    "directory": {
-      "type": "string",
-      "description": "A directory where the project is placed"
-    }
-  },
-  "required": [
-    "name"
-  ]
-}
