+++ libs/webcomponents-nx/src/index.ts
@@ -1,0 +1,1 @@
+export * from './executors/sync/executor';
